// src/components/Questionnaire/enums/questionnaireEnums.ts
export enum Reason {
    escapeFullTimeJob = "Replace my full time job",
    passiveIncome = "Build a semi-passive income stream",
    expandPortfolio = "Expand my business portfolio",
    personalPassion = "Pursue a passion or personal interest"
}

export enum Vision {
    singleLocation = "Operate a single location successfully",
    multiLocation = "Build multiple locations in my region",
    growthFocused = "Grow and sell the business within 5 - 10 years",
    legacy = "Create a family legacy business"
}

export enum Involvement {
    fullTime = "Full time owner-operator",
    partTime = "Part time / semi-absentee",
    investor = "Investor only (hire management team)"
}


export enum Capital {
    smallCap = "Under $100,000",
    medCap = "$150,000 - $500,000",
    largeCap =  "$550,000 - $1,000,000",
    megaCap = "Over $1,000,000"
}